Third commit, download and e-mail functionality added, Oauth2 security in development

+++ src/main/java/com/example/school_diary_end_project/controllers/PupilController.java
@@ -1,16 +1,14 @@
 package com.example.school_diary_end_project.controllers;
 
 import com.example.school_diary_end_project.controllers.util.RESTError;
+import com.example.school_diary_end_project.entities.DepartmentEntity;
 import com.example.school_diary_end_project.entities.PupilEntity;
 import com.example.school_diary_end_project.entities.enums.EUserRole;
-import com.example.school_diary_end_project.repositories.ParentRepository;
-import com.example.school_diary_end_project.repositories.PupilRepository;
+import com.example.school_diary_end_project.repositories.*;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -27,14 +25,17 @@ public class PupilController {
     @Autowired
     private PupilRepository pupilRepo;
 
-/*    @Autowired
-    private SubjectRepository subRepo;*/
+    @Autowired
+    private SubjectRepository subRepo;
 
     @Autowired
     private ParentRepository parentRepo;
 
-  /*  @Autowired
-    private DepartmentRepository depoRepo;*/
+    @Autowired
+    private DepartmentRepository depoRepo;
+
+    @Autowired
+    private GradeRepository gradeRepo;
 
     @RequestMapping
     public ResponseEntity<?> getDb() {
@@ -80,14 +81,6 @@ public class PupilController {
         return new ResponseEntity<RESTError>(new RESTError(1, "Pupil not found"), HttpStatus.NOT_FOUND);
     }
 
-  /*  // Opciona metoda za testiranje
-    @RequestMapping(method = RequestMethod.PUT, value = "/{id}/subject/{ids}")
-    public ResponseEntity<?> editSubject(@PathVariable Integer id, @PathVariable Integer ids) {
-        SubjectEntity temp = subRepo.findById(ids).get();
-        pupilRepo.findById(id).get().addSubjects(temp);
-        return new ResponseEntity<PupilEntity>(pupilRepo.save(pupilRepo.findById(id).get()), HttpStatus.OK);
-
-    }*/
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<?> deletePupil(@PathVariable Integer id) {
@@ -96,9 +89,31 @@ public class PupilController {
             PupilEntity temp = pupilRepo.findById(id).get();
             String exit = String.format("Pupil %s %s has been deleted",temp.getName(), temp.getSurname());
 //            Deletes a parent if a parent has only one child
-            if (pupilRepo.findAllByParent(temp.getParent()).size() == 1){
-                parentRepo.delete(temp.getParent());
+
+            if (temp.getParent() !=null){
+                if (pupilRepo.findAllByParent(temp.getParent()).size() == 1){
+                    parentRepo.delete(temp.getParent());
+                }
+
+                if (pupilRepo.findAllByParent(temp.getParent()).size() > 1){
+                    parentRepo.findById(temp.getDepartment().getId()).get().getChildren().remove(temp);
+                    parentRepo.save(temp.getParent());
+
+                }
             }
+
+            if (temp.getDepartment() != null){
+                depoRepo.findById(temp.getDepartment().getId()).get().getPupils().remove(temp);
+                depoRepo.save(temp.getDepartment());
+
+            }
+
+            if (temp.getGrades().size() > 0){
+                gradeRepo.deleteAllByPupil(temp);
+            }
+
+
+
 
 
 
@@ -121,40 +136,42 @@ public class PupilController {
         return new ResponseEntity<RESTError>(new RESTError(1, "Pupil or Parent not found"), HttpStatus.NOT_FOUND);
     }
 
- /*   @RequestMapping(value = "/{id}/setDepartment/{idd}", method = RequestMethod.PUT)
+    @RequestMapping(value = "/{id}/removeParent", method = RequestMethod.PUT)
+    public ResponseEntity<?> removeParent(@PathVariable Integer id) {
+        if (pupilRepo.findById(id).isPresent()) {
+            pupilRepo.findById(id).get().setParent(null);
+            return new ResponseEntity<PupilEntity>(pupilRepo.save(pupilRepo.findById(id).get()), HttpStatus.OK);
+
+        }
+        return new ResponseEntity<RESTError>(new RESTError(1, "Pupil not found"), HttpStatus.NOT_FOUND);
+    }
+
+    @RequestMapping("/{id}")
+    public ResponseEntity<?> findById(@PathVariable Integer id){
+        if (pupilRepo.findById(id).isPresent()){
+            return new ResponseEntity<PupilEntity>(pupilRepo.findById(id).get(), HttpStatus.OK);
+        }
+        return new ResponseEntity<RESTError>(new RESTError(1, "Pupil not found"), HttpStatus.NOT_FOUND);
+    }
+
+    @RequestMapping(value = "/{id}/setDepartment/{idd}", method = RequestMethod.PUT)
     public ResponseEntity<?> setDepartment(@PathVariable Integer id, @PathVariable Integer idd) {
         if (pupilRepo.findById(id).isPresent() && depoRepo.findById(idd).isPresent()) {
             pupilRepo.findById(id).get().setDepartment(depoRepo.findById(idd).get());
-            if (!depoRepo.findById(idd).get().getSchoolYear().getSubjectsListened().isEmpty()) {
-                Set<SubjectEntity> subjectsSet = new HashSet<>(depoRepo.findById(idd).get().getSchoolYear().getSubjectsListened());
-                pupilRepo.findById(id).get().setSubjects(subjectsSet);
-            }
-
-
             return new ResponseEntity<PupilEntity>(pupilRepo.save(pupilRepo.findById(id).get()), HttpStatus.OK);
         }
         return new ResponseEntity<RESTError>(new RESTError(1, "Pupil or Department not found"), HttpStatus.NOT_FOUND);
 
-    }*/
+    }
 
-  /*  //azurira listu predmeta za ucenike(izmeniti)
-    public ResponseEntity<?> updateSubjects(@PathVariable Integer id) {
-        if (pupilRepo.findById(id).isPresent() && pupilRepo.findById(id).get().getDepartment().equals(null)) {
-            //bezveze
-        }
+    @RequestMapping(value = "/{id}/removeDepartment", method = RequestMethod.PUT)
+    public ResponseEntity<?> removeDepartment(@PathVariable Integer id) {
         if (pupilRepo.findById(id).isPresent()) {
-            PupilEntity temp = pupilRepo.findById(id).get();
-            if (temp.getDepartment().equals(null) && temp.getDepartment().getSchoolYear().equals(null)
-                    && temp.getDepartment().getSchoolYear().getSubjectsListened().isEmpty()) {
-
-                return new ResponseEntity<RESTError>(new RESTError(1, "No subjects found"), HttpStatus.NOT_FOUND);
-
-            }
-            temp.setSubjects((Set<SubjectEntity>) temp.getDepartment().getSchoolYear().getSubjectsListened());
-
+            pupilRepo.findById(id).get().setDepartment(null);
             return new ResponseEntity<PupilEntity>(pupilRepo.save(pupilRepo.findById(id).get()), HttpStatus.OK);
         }
-        return new ResponseEntity<RESTError>(new RESTError(1, "Pupil or Subjects not found"), HttpStatus.NOT_FOUND);
+        return new ResponseEntity<RESTError>(new RESTError(1, "Pupil not found"), HttpStatus.NOT_FOUND);
+    }
 
-    }*/
+
 }
