+++ src/main/java/com/example/school_diary_end_project/controllers/SubjectController.java
@@ -2,12 +2,15 @@ package com.example.school_diary_end_project.controllers;
 
 import com.example.school_diary_end_project.controllers.util.RESTError;
 import com.example.school_diary_end_project.entities.SubjectEntity;
+import com.example.school_diary_end_project.repositories.GradeRepository;
+import com.example.school_diary_end_project.repositories.ScheduleRepository;
 import com.example.school_diary_end_project.repositories.SubjectRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.persistence.criteria.CriteriaBuilder;
 import java.util.List;
 
 @RestController
@@ -17,6 +20,11 @@ public class SubjectController {
     @Autowired
     private SubjectRepository subjectRepo;
 
+    @Autowired
+    private GradeRepository gradeRepo;
+
+    @Autowired
+    private ScheduleRepository scheduleRepo;
 
 
     @RequestMapping
@@ -46,11 +54,11 @@ public class SubjectController {
                 temp.setSemesterlessonsfund(subject.getSemesterlessonsfund());
             }
 
-            if (!subject.getDescription().equals(null)){
+            if (!subject.getDescription().equals(null)) {
                 temp.setDescription(subject.getDescription());
             }
 
-            if (!subject.getYear().equals(null)){
+            if (!subject.getYear().equals(null)) {
                 temp.setYear(subject.getYear());
             }
             return new ResponseEntity<SubjectEntity>(subjectRepo.save(subjectRepo.findById(id).get()), HttpStatus.OK);
@@ -60,13 +68,25 @@ public class SubjectController {
 
     }
 
+
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<?> deleteSubject(@PathVariable Integer id) {
         if (subjectRepo.findById(id).isPresent()) {
             SubjectEntity temp = subjectRepo.findById(id).get();
+            if (temp.getGrades().size() > 0) {
+                return new ResponseEntity<RESTError>(new RESTError(1, String.format
+                        ("Subject %s cannot be deleted, remove its references to all grades beforehand", temp.getName())),
+                        HttpStatus.BAD_REQUEST);
+            }
+
+            if (temp.getSchedule().size() > 0) {
+                return new ResponseEntity<RESTError>(new RESTError(1, String.format
+                        ("Subject %s cannot be deleted, remove its references to all schedule instances beforehand"
+                                , temp.getName())), HttpStatus.BAD_REQUEST);
+            }
 
             subjectRepo.deleteById(id);
-            return new ResponseEntity<String>(String.format("Subject %s has been deleted",temp.getName()), HttpStatus.OK);
+            return new ResponseEntity<String>(String.format("Subject %s has been deleted", temp.getName()), HttpStatus.OK);
         }
         return new ResponseEntity<RESTError>(new RESTError(1, "Subject not found"), HttpStatus.NOT_FOUND);
     }
@@ -82,25 +102,24 @@ public class SubjectController {
     }
 
 
+    @RequestMapping(method = RequestMethod.PUT, value = "/{id}/removeReferences")
+    public ResponseEntity<?> removeReferences(@PathVariable Integer id){
+        if (subjectRepo.findById(id).isPresent()){
+            SubjectEntity temp = subjectRepo.findById(id).get();
 
-    // dodeljuje predmete razredima za koji su predvidjeni (1-4 razred)
-//    @RequestMapping(method = RequestMethod.PUT, value = "/update")
-//    public ResponseEntity<?> updateSubjectYear() {
-//        List<SubjectEntity> lista = (List<SubjectEntity>) subjectRepo.findAll();
-//        for (SubjectEntity subject : lista) {
-//            if (subject.getPlanForYear().equals(null)) {
-//                continue;
-//            }
-//
-//            if (subject.getPlanForYear().equals(classRepo.findByYear(subject.getPlanForYear()).getYear())) {
-//                subject.setClassYear(classRepo.findByYear(subject.getPlanForYear()));
-//            }
-//            subjectRepo.save(subject);
-//
-//        }
-//
-//        subjectRepo.saveAll(lista);
-//
-//        return new ResponseEntity<List<SubjectEntity>>((List<SubjectEntity>) subjectRepo.saveAll(lista), HttpStatus.OK);
-//    }
+            if (temp.getGrades().size() > 0 ){
+                gradeRepo.deleteAllBySubject(temp);
+            }
+
+            if (temp.getSchedule().size()>0){
+                scheduleRepo.deleteAllBySubject(temp);
+            }
+
+            return new ResponseEntity<SubjectEntity>(subjectRepo.findById(id).get(), HttpStatus.OK);
+        }
+
+        return new ResponseEntity<RESTError>(new RESTError(1, "Subject not found"), HttpStatus.NOT_FOUND);
+    }
+
+
 }
