+++ src/main/java/com/example/school_diary_end_project/controllers/TeacherController.java
@@ -3,6 +3,10 @@ package com.example.school_diary_end_project.controllers;
 
 import com.example.school_diary_end_project.controllers.util.RESTError;
 import com.example.school_diary_end_project.entities.TeacherEntity;
+import com.example.school_diary_end_project.entities.enums.EUserRole;
+import com.example.school_diary_end_project.repositories.DepartmentRepository;
+import com.example.school_diary_end_project.repositories.GradeRepository;
+import com.example.school_diary_end_project.repositories.ScheduleRepository;
 import com.example.school_diary_end_project.repositories.TeacherRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,25 +22,30 @@ public class TeacherController {
     @Autowired
     private TeacherRepository teachRepo;
 
-/*    @Autowired
-    private SubjectRepository subRepo;*/
 
-//    @Autowired
-//    private DepartmentRepository depoRepo;
+    @Autowired
+    private DepartmentRepository depoRepo;
+
+    @Autowired
+    private GradeRepository gradeRepo;
+
+    @Autowired
+    private ScheduleRepository scheduleRepo;
 
     @RequestMapping
-    public ResponseEntity<?> getDb(){
-        return new ResponseEntity<List<TeacherEntity>>((List<TeacherEntity>)teachRepo.findAll(), HttpStatus.OK);
+    public ResponseEntity<?> getDb() {
+        return new ResponseEntity<List<TeacherEntity>>((List<TeacherEntity>) teachRepo.findAll(), HttpStatus.OK);
     }
 
 
     @RequestMapping(method = RequestMethod.POST)
-    public ResponseEntity<?> addTeacher(@RequestBody TeacherEntity teacher){
+    public ResponseEntity<?> addTeacher(@RequestBody TeacherEntity teacher) {
+        teacher.setRole(EUserRole.ROLE_TEACHER);
         return new ResponseEntity<TeacherEntity>(teachRepo.save(teacher), HttpStatus.OK);
     }
 
     @RequestMapping(method = RequestMethod.PUT, value = "/{id}")
-    public ResponseEntity<?> editTeacher(@RequestBody TeacherEntity teacher, @PathVariable Integer id){
+    public ResponseEntity<?> editTeacher(@RequestBody TeacherEntity teacher, @PathVariable Integer id) {
         if (teachRepo.findById(id).isPresent()) {
             TeacherEntity temp = teachRepo.findById(id).get();
 
@@ -64,11 +73,11 @@ public class TeacherController {
                 temp.setUsername(teacher.getUsername());
             }
 
-            if (!teacher.getTeacherType().equals(null)) {
+            if ( teacher.getTeacherType() != null) {
                 temp.setTeacherType(teacher.getTeacherType());
             }
 
-            if (!teacher.getQualifiedForSubjects().equals(null)){
+            if (!teacher.getQualifiedForSubjects().equals(null)) {
                 temp.setQualifiedForSubjects(teacher.getQualifiedForSubjects());
             }
 
@@ -80,10 +89,25 @@ public class TeacherController {
     }
 
     @RequestMapping(method = RequestMethod.DELETE, value = "/{id}")
-    public ResponseEntity<?> deleteTeacher(@PathVariable Integer id){
+    public ResponseEntity<?> deleteTeacher(@PathVariable Integer id) {
         if (teachRepo.findById(id).isPresent()) {
             TeacherEntity temp = teachRepo.findById(id).get();
-            String exit = String.format("Teacher %s %s has been deleted",temp.getName(), temp.getSurname());
+            String exit = String.format("Teacher %s %s has been deleted", temp.getName(), temp.getSurname());
+            if (temp.getGrades().size() > 0) {
+                return new ResponseEntity<RESTError>(new RESTError(1, "Teacher cannot be deleted, please remove his connections to grades first"), HttpStatus.BAD_REQUEST);
+
+
+            }
+
+            if (temp.getSchedule().size() > 0) {
+                return new ResponseEntity<RESTError>(new RESTError(1, "Teacher cannot be deleted, please remove his connections to schedule first"), HttpStatus.BAD_REQUEST);
+            }
+
+            if (temp.getHeadOfDepartment() != null) {
+                return new ResponseEntity<RESTError>(new RESTError(1, "Teacher cannot be deleted, please remove his connections as head of department first"), HttpStatus.BAD_REQUEST);
+            }
+
+
             teachRepo.deleteById(id);
 
             return new ResponseEntity<String>(exit, HttpStatus.OK);
@@ -94,21 +118,15 @@ public class TeacherController {
 
     }
 
-  /*  //	dodeljuje predmet profesoru
-    @RequestMapping(method = RequestMethod.PUT, value = "/{id}/subject/{ids}")
-    public ResponseEntity<?> assignSubject(@PathVariable Integer id, @PathVariable Integer ids){
-        if (teachRepo.findById(id).isPresent() && subRepo.findById(ids).isPresent()) {
-
-            Set<SubjectEntity> subjects = teachRepo.findById(id).get().getSubjects();
-            subjects.add(subRepo.findById(ids).get());
-
-            teachRepo.findById(id).get().setSubjects(subjects);
-
-            return new ResponseEntity<TeacherEntity>(teachRepo.save(teachRepo.findById(id).get()), HttpStatus.OK);
+    @RequestMapping("/{id}")
+    private ResponseEntity<?> findById(@PathVariable Integer id){
+        if (teachRepo.findById(id).isPresent()){
+            return new ResponseEntity<TeacherEntity>(teachRepo.findById(id).get(), HttpStatus.OK);
         }
-
-        return new ResponseEntity<RESTError>(new RESTError(1, "Teacher or Subject not found"), HttpStatus.NOT_FOUND);
+        return new ResponseEntity<RESTError>(new RESTError(1, "Teacher not found"), HttpStatus.NOT_FOUND);
     }
+
+
 
     //	dodeljuje razrednog staresinu odeljenju
     @RequestMapping(value = "/{id}/department/{idd}", method = RequestMethod.PUT)
@@ -120,5 +138,48 @@ public class TeacherController {
         }
 
         return new ResponseEntity<RESTError>(new RESTError(1, "Teacher or Department not found"), HttpStatus.NOT_FOUND);
-    }*/
+    }
+
+
+    //	oduzima staresinstvo razrednom
+    @RequestMapping(value = "/{id}/removeDepartment/{idd}", method = RequestMethod.PUT)
+    public ResponseEntity<?> removeHeadOfDepartment(@PathVariable Integer id, @PathVariable Integer idd){
+        if (teachRepo.findById(id).isPresent() && teachRepo.findById(id).get().getHeadOfDepartment() != null) {
+            depoRepo.findById(idd).get().setHeadTeacher(null);
+            depoRepo.save(depoRepo.findById(id).get());
+            teachRepo.findById(id).get().setHeadOfDepartment(null);
+
+            return new ResponseEntity<TeacherEntity>(teachRepo.save(teachRepo.findById(id).get()), HttpStatus.OK);
+        }
+
+        return new ResponseEntity<RESTError>(new RESTError(1, "Teacher or Department not assigned found"), HttpStatus.NOT_FOUND);
+    }
+
+
+
+
+// Removes all assigned references
+    @RequestMapping(method = RequestMethod.PUT, value = "/{id}/unAssign")
+    public ResponseEntity<?> removeAllReferences(@PathVariable Integer id){
+        if(teachRepo.findById(id).isPresent()){
+            TeacherEntity temp = teachRepo.findById(id).get();
+            if (temp.getHeadOfDepartment() != null){
+                depoRepo.findById(temp.getHeadOfDepartment().getId()).get().setHeadTeacher(null);
+                depoRepo.save(depoRepo.findById(temp.getHeadOfDepartment().getId()).get());
+            }
+
+            if (temp.getGrades().size() > 0){
+                gradeRepo.deleteAllByTeacher(temp);
+            }
+
+            if (temp.getSchedule().size() > 0){
+                scheduleRepo.deleteAllByTeacher(temp);
+            }
+
+            return new ResponseEntity<TeacherEntity>(teachRepo.save(teachRepo.findById(id).get()), HttpStatus.OK);
+        }
+        return new ResponseEntity<RESTError>(new RESTError(1, "Teacher not found"), HttpStatus.NOT_FOUND);
+
+    }
+
 }
